refactor: made MonthlySpendingLimit reusable by creating two new components set and display limit
